refactor to wiseloculus api
