+++ wiseloculus.py
@@ -1,0 +1,49 @@
+"""Implements the Wiseloculus API Queries."""
+
+import logging
+import aiohttp
+import asyncio
+import yaml
+
+# Load configuration from config.yaml
+with open('config.yaml', 'r') as file:
+    config = yaml.safe_load(file)
+
+
+server_ip = config.get('server', {}).get('lapis_address', 'http://default_ip:8000')
+
+async def fetch_data(session, mutation, date_range):
+    payload = {
+        "aminoAcidMutations": [mutation],
+        "sampling_dateFrom": date_range[0].strftime('%Y-%m-%d'),
+        "sampling_dateTo": date_range[1].strftime('%Y-%m-%d'),
+        "fields": ["sampling_date"]
+    }
+
+    async with session.post(
+        f'{server_ip}/sample/aggregated',
+        headers={
+            'accept': 'application/json',
+            'Content-Type': 'application/json'
+        },
+        json=payload
+    ) as response:
+        if response.status == 200:
+            data = await response.json()
+            return {"mutation": mutation,
+                    "data": data.get('data', [])}
+        else:
+            logging.error(f"Failed to fetch data for mutation {mutation}.")
+            logging.error(f"Status code: {response.status}")
+            logging.error(await response.text())
+            return {"mutation": mutation,
+                    "data": None}
+        
+async def fetch_single_mutation(mutation, date_range):
+    async with aiohttp.ClientSession() as session:
+        return await fetch_data(session, mutation, date_range)
+
+async def fetch_all_data(mutations, date_range):
+    async with aiohttp.ClientSession() as session:
+        tasks = [fetch_data(session, mutation, date_range) for mutation in mutations]
+        return await asyncio.gather(*tasks)
